--- a/internal/gen/cldr.go.tmpl
+++ b/internal/gen/cldr.go.tmpl
@@ -20,6 +20,8 @@ var calendarMonthNames = [...]calendarMonths{
 {{- end }}
 }
 
+// monthIndexes contains indexes of months names, each calendar has 6 indexes
+// for all variations of "width" and "context".
 type monthIndexes [18]int16
 
 var monthLookup = map[string]monthIndexes{
@@ -28,19 +30,10 @@ var monthLookup = map[string]monthIndexes{
 {{- end }}
 }
 
-func fmtMonth(locale, calendar, context, width string) func(int, string) string {
+func fmtMonth(locale string, calendar calendarType, context, width string) func(int, string) string {
   indexes := monthLookup[locale]
 
-  var i int
-
-  switch calendar {
-  case "gregorian": // noop
-    // "gregorian" calendar index is 0
-  case "buddhist":
-    i = 6 // 1*3*2
-  case "persian":
-    i = 12 // 2*3*2
-  }
+  i := calendar*6
 
   // "abbreviated" width index is 0
   switch width {
@@ -102,14 +95,14 @@ func defaultNumberingSystem(locale language.Tag) numberingSystem {
   }
 }
 
-func defaultCalendar(locale language.Tag) string {
+func defaultCalendar(locale language.Tag) calendarType {
   switch v, _ := locale.Region(); v.String() {
   default:
-    return "gregorian"
+    return calendarTypeGregorian
 {{- range $v := .CalendarPreferences }}
   {{- if ne (index $v.Calendars 0) "gregorian" }}
   case "{{ join $v.Regions "\", \"" }}":
-    return "{{ index $v.Calendars 0 }}"
+    return calendarType{{ title (index $v.Calendars 0) }}
   {{- end }}
 {{- end }}
   }