```console
$ benchstat a.txt b.txt
goos: darwin
goarch: arm64
pkg: go.expect.digital/intl
cpu: Apple M2 Max
                         │    a.txt    │                b.txt                │
                         │   sec/op    │   sec/op     vs base                │
NewDateTime/fa-IR-12       231.0n ± 2%   207.9n ± 2%  -10.00% (p=0.000 n=10)
NewDateTime/lv-LV-12       295.9n ± 1%   276.8n ± 1%   -6.45% (p=0.000 n=10)
NewDateTime/dz-BT-12       299.3n ± 1%   281.1n ± 1%   -6.10% (p=0.000 n=10)
DateTime_Format/fa-IR-12   828.7n ± 0%   824.6n ± 0%   -0.49% (p=0.022 n=10)
DateTime_Format/lv-LV-12   190.1n ± 1%   188.3n ± 0%   -0.97% (p=0.000 n=10)
DateTime_Format/dz-BT-12   473.9n ± 1%   468.3n ± 2%   -1.18% (p=0.005 n=10)
geomean                    339.4n        324.9n        -4.27%

                         │   a.txt    │                b.txt                 │
                         │    B/op    │    B/op     vs base                  │
NewDateTime/fa-IR-12       309.0 ± 0%   269.0 ± 0%  -12.94% (p=0.000 n=10)
NewDateTime/lv-LV-12       277.0 ± 0%   237.0 ± 0%  -14.44% (p=0.000 n=10)
NewDateTime/dz-BT-12       277.0 ± 0%   237.0 ± 0%  -14.44% (p=0.000 n=10)
DateTime_Format/fa-IR-12   104.0 ± 0%   104.0 ± 0%        ~ (p=1.000 n=10) ¹
DateTime_Format/lv-LV-12   24.00 ± 0%   24.00 ± 0%        ~ (p=1.000 n=10) ¹
DateTime_Format/dz-BT-12   144.0 ± 0%   144.0 ± 0%        ~ (p=1.000 n=10) ¹
geomean                    142.9        132.6        -7.23%
¹ all samples are equal

                         │   a.txt    │                b.txt                │
                         │ allocs/op  │ allocs/op   vs base                 │
NewDateTime/fa-IR-12       5.000 ± 0%   5.000 ± 0%       ~ (p=1.000 n=10) ¹
NewDateTime/lv-LV-12       6.000 ± 0%   6.000 ± 0%       ~ (p=1.000 n=10) ¹
NewDateTime/dz-BT-12       6.000 ± 0%   6.000 ± 0%       ~ (p=1.000 n=10) ¹
DateTime_Format/fa-IR-12   23.00 ± 0%   23.00 ± 0%       ~ (p=1.000 n=10) ¹
DateTime_Format/lv-LV-12   4.000 ± 0%   4.000 ± 0%       ~ (p=1.000 n=10) ¹
DateTime_Format/dz-BT-12   24.00 ± 0%   24.00 ± 0%       ~ (p=1.000 n=10) ¹
geomean                    8.575        8.575       +0.00%
¹ all samples are equal
```